pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.13",
    "discord-py>=2.5.2",
    "feedparser>=6.0.11",
    "flask>=3.1.1",
]
